BACKEND/PROJECT/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.5
django-allauth==0.51.0
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-filter==22.1
django-js-asset==2.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
oauthlib==3.2.0
Pillow==9.1.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
urllib3==1.26.10
